renames <Groups/> and GroupsContainer to GroupManagement and moves DataPanel dir inside GroupManagement folder
