+++ src/login/components/GroupManagement/DataPanel/DataPanel.js
@@ -1,8 +1,8 @@
 import React, { Component, Fragment } from "react";
 import PropTypes from "prop-types";
-import InjectIntl from "../../translation/InjectIntl_HOC_factory";
-import GroupList from "./GroupList";
-import EditMode from "./EditMode";
+import InjectIntl from "../../../translation/InjectIntl_HOC_factory";
+import GroupList from "../GroupList";
+import EditMode from "../EditMode/EditMode";
 
 class DataPanel extends Component {
   state = { editMode: false, group: "" };
