+++ src/login/components/GroupManagement/Wizard/WizardPanel.js
@@ -1,6 +1,6 @@
 import React, { Component, Fragment } from "react";
 import PropTypes from "prop-types";
-import InjectIntl from "../../translation/InjectIntl_HOC_factory";
+import InjectIntl from "../../../translation/InjectIntl_HOC_factory";
 
 const RenderCreateGroupWizard = (props) => {
   return (
